Add test for mocking API call
This is a working implementation of the API being mocked out.
However, there is still a bug where the onCreate() method is called
twice, the first time it's invoking https://quotes.rest/qod rather
than the mocked response. The second time it is invoked, the mocked
response is be called.