Testing main activity
I'm in the middle of a test right now, but I'm having some trouble
because I can't mock a final class/method. I'm trying to figure out
how I can go about testing this.

For context...

I have an AsyncTask that calls out to an API, that call is in the
onCreate() method of the MainActivity. I want to verify that if the
call has already been made, and the user is coming back to the app
after pausing that a subsequent API call is not made. Hence, I want
to verify that the execute() method was not called on the AsyncTask.